renamed apiConstants file
