+++ app/views/SymfoniViews/SymfoniRecieveVPView.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { View, Text, ActivityIndicator } from 'react-native';
 import QRCode from 'react-native-qrcode-svg';
 import { TYPE_PERSON_CREDENTIAL } from '../../../src/constants/verifiableCredentialConstants';
-import { SYMFONI_GET_MAIN_IDENTIFIER_URL } from '../../apiConstants';
+import { SYMFONI_GET_MAIN_IDENTIFIER_URL } from '../../api.constants.';
 import { IMainIdentifier } from '../../interfaces/IMainIdentifier.interface';
 import { styles } from '../../styles';
